Migrating error checking methods
